refactor client logic to filament package
